+++ CHANGELOG.rst
@@ -1,6 +1,6 @@
 In development
 --------------
-* Color: Added color module, support for color added to cli
+* Color: Added colors module, support for colors added to cli
 * Machines: Added ``.get()`` method for checking several commands.
 
 1.5.0